snapuserd: Add a userspace implementation of IBlockServer etc.
This implements IBlockServer entirely in userspace. Effectively, it is
dm-user without the kernel module. Of course, we can't actually create
block devices this way, so there is no way to perform syscalls through
this implementation. However, it provides userspace methods to do the
same thing as read(2), and snapuserd_tests have been refactored to use
abstraction layers around syscalls.

Bug: 288273605
Test: snapuserd_test
Change-Id: Id8c0cd0f9857041ff79dfe1f93d5072be180f083