Revert "Only allow debuggable Microdroid VMs to mount /vendor"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a02ccd3b-b1a1-46f0-88b1-fc9893c78ac5). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I0c49324bda6f9e03ae423d1a3986c6259ef9d657