x86: mmap rnd bits: Allow min of 30
With a regular x86 kernel we are always able to get 32
min bits for mmap randomization.

However to emulate 16KB for x86 app developers the kernel
can only provide 30 bits for the randomization due the to
larger page size (PAGE_SHIFT =+ 2).

Allow the set SetMmapRndBitsMin() to fallback until 30 bits.

This should be a no-op change for all devices, except the
x86 16KB emulator, since they will always be able to get
32 random bits. For the emulated case gracefully retry
until 30 bits.

Bug: 309816695
Test: Boot test 4k x86 device
Test: Boot test emulated 16k x86 device
Change-Id: I48d47d56ac3aecb71a9e0093a7033bb60b89c2b7
Signed-off-by: Kalesh Singh <kaleshsingh@google.com>