init.rc: stop calling 'fsverity_init --load-verified-keys'
Since Android 14, Android does not use fsverity builtin signatures.
(fsverity remains supported, but signatures are verified in userspace,
or it is used for integrity-only.)  Therefore, the only reason to still
run 'fsverity_init --load-verified-keys' at boot time is to ensure that
old files can still be opened, if:

- They were created by Android 13 or earlier, with an fsverity builtin
  signature by a key in /{system,product}/etc/security/fsverity/.

- *And*, the kernel still has CONFIG_FS_VERITY_BUILTIN_SIGNATURES=y.
  (Disabling it avoids the problem, as files with builtin signatures
  will then be openable even if the signing key is not loaded.)

However, it appears that this won't actually be relevant for any
devices.  Therefore, remove the corresponding code.

Bug: 290064770
Test: presubmit
Change-Id: I3376c3f0b4b9bd4ba2fd614259522be0c1daafb6