Revert "Remove local_include_dirs from fs_mgr_defaults."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/6d1f425d-0ff2-4be6-bdfa-5bce548cb790). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I68c0b6d3b6d329bdfd167fcb098573190856919e