libprocessgroup: optimize SetAttributeAction::ExecuteForProcess performance
Current implementation of SetAttributeAction::ExecuteForProcess reuses
SetAttributeAction::ExecuteForTask while not utilizing available uid/pid
information. This results in a call to GetPathForTask() which is an
expensive function due to it reading and parsing /proc/$pid/cgroups.
This can be avoided if we provide cgroup path template in the
task_profiles.json and utilize available uid/pid info.
Add new Path attribute to Attributes in task_profiles.json for attribute
path template and use uid/pid in SetAttributeAction::ExecuteForProcess
to generate path for the attribute.

Bug: 292636609
Change-Id: I02e3046bd85d0dfebc68ab444f1796bb54cc69c7
Signed-off-by: Suren Baghdasaryan <surenb@google.com>