Revert "init.rc: stop calling 'fsverity_init --load-verified-keys'"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/8580317e-d6a3-480f-96ed-172c2061ea89). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Icba38d4fa46ee50f75a8ab55273f8c90cd83f153