Revert "Use /bootstrap-apex for bootstrap APEXes"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a4d84082-d35f-4b2b-8f9c-61159010f979). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I38a3d49387dcd8151647feb7c3e68cf74c856427