Revert "Use /bootstrap-apex for bootstrap APEXes"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/aec73b48-0156-4581-b6cb-eb5b80422685). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iaead932cd282b94119294c3f1edd24b12bf76213