Revert "Use /bootstrap-apex for bootstrap APEXes"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/4eaa9635-f4d2-4ce4-acdf-19f726beae52). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I0dc0909c420f72dc4ea1f364ed0ed41b31ecafd1