Revert "Use /bootstrap-apex for bootstrap APEXes"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/acb3fa1f-b02e-4e77-ba33-a4f811b86b9e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I8c2e3fda5c6b37b3031185ea52ed34a530363bc2