Revert "Use /bootstrap-apex for bootstrap APEXes"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b7dc8d1b-7861-4410-afb2-ccccf099c04a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I6cbbfe827dbca96475c8e2363cb45f8860ecbdc9