Revert "Use /bootstrap-apex for bootstrap APEXes"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/dd65b239-0945-486e-8404-2163c232d175). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia42f38ce38310e5f10549da6a4fe5b0d83d27ee8