Revert "Make libdebuggerd ramdisk available."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/3111a2b3-ed0f-45b8-b949-14ce7f7ae766). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I9fef3888d35809bc448e5101a5373f0d7782a6c5