Revert "Use /bootstrap-apex for bootstrap APEXes"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1616589c-b0fb-4fe5-8610-076e2f368028). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I739e68034b48f6d29f922476476bd8a7312ca07f