Revert "libsnapshot: Introduce a dynamically linked version of l..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/6936b23f-fa6f-4ecd-9a39-f7aa51d2e3dc). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iddba0e2d7c9a2e26ec315a313d25c1ecdd19125c