Revert "Read .rc files from bootstrap apexes"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d1ad5d5f-9b26-4b9f-8f0b-d47a7cea496a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id3a7944d337d55fca5b8d83b897f5ebe62339745