Revert "Remove cutils threads.h header completely."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/63e093cf-5f6b-4ec0-835f-ff67310e6cba). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I7e62997c5275c889700c1f6252ebf4c274919d9e