Add unit test for cow compressor performance
Adding test to measure performance differences between cow compression
algorithms + levels. Ultimately we want this to be separate from
cow_api_test so it would be nice to have this be it's own binary. Can
add some tests for decompression + compressing from a part of an actual
img file too.

Test: m cow_compressor_test
Change-Id: Iba92ae3c0b2ad4ff6f842556b701b223d7d37823