[RFC]fastboot: Make `fastboot devices` output values if serial device is set to a network device
This makes it so if the serial value is set via the -s argument,
or via ANDROID_SERIAL or FASTBOOT_DEVICE env variables, the
command `fastboot devices` won't return empty.

This might be useful for helping with automation like tradefed
which which uses this command to detect when devices appear.

Change-Id: If4db6fc287cd2f65173a109b6515c20424597b97
Signed-off-by: John Stultz <jstultz@google.com>