Revert "Export create_snapshot to sdk targets"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f3e96291-d09c-483a-b74a-c1a4590fe9af). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I861bd38668644bf53bc1df091301b85c4dcebf14