Revert "Drop all path-related methods from android::String8"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1b718a46-ade7-4677-8422-e88bae6b404d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I2adb309e777666ad7da4a107c065599ffab29aa4