Revert "Drop all path-related methods from android::String8"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/961ec978-c7e5-4231-94d5-96ac4b3bc9a3). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia034d309e0f42ce6db96417b5fceff420d35053c