Revert "Revert "Revert "Drop all path-related methods from andro..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e23d67da-8153-4ee0-a54f-9c5afbc4a3d3). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I8be7b9b724ca5ac9d2d5aa0bfeec60fed4e19385