Revert "Revert "Revert "Drop all path-related methods from andro..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/db517be0-6e62-471c-9532-cb3724082d77). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I15b33716fe3e3ab1813e40a01d2cb576a7284672