Revert "Revert "Revert "Drop all path-related methods from andro..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/37ccdcdf-c101-477b-9b85-b967f3ba5a75). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I17bd3dd2bffc3a838e0938bff3177a763ac7737f