Revert "Revert "Revert "Drop all path-related methods from andro..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ddc43fce-aeaf-4ced-b4c7-f3c763621f0c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I824643d10136c1cb1f08315ceb95b4526073742d