Adding compression to cow header
Adding Version 2 of CowHeader. This will inherit from original
CowHeader and add a compression field. We are no longer supporting
different compressions per operation so having this one field is enough.
Also need to change COW reader translate the header from on disk to new
format.

note-> not sure how this got on this CL, but we are also updating
BLOCK_SZ name to block_size as we should reserve all caps for compile
time constants

Test: cow_api_test
Change-Id: If88dceda139807cc5e647b706ddeb2b3e83c024f