Revert "Lock down String8|16.string() usage"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d8d402a3-1609-4252-a14e-2742f3f9f8d1). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I2c5525d3939b8295c6b2acb1d4d4824e0ca4fd43