Revert "Lock down String8|16.string() usage"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/47c59677-2528-4e09-b05f-a04c71ff7200). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Idbf184e87b483f3fefa0eeadb30e141984f077e1