Retrofit device secondary partition flashing
Secondary partitions on retrofit devices should be flashed in the
bootloader as they are physical partitions. Adding this check to
Addtasks so that the partitions are flashed in the correct place.
Removing the delete task as this never worked before. is_logical for
secondary partitions will return as false on retrofit_device (in
fastbootd).

Test: fastboot update on Sargo device
Change-Id: Ia0870240b636df398d5bf004073086a7dc62021f