Revert "Only skip _16k/_64k kernel modules dirs on 4K builds"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/7372d847-01e0-41d5-aedb-4a6380f9f644). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I75e67ade4206136b97135103b47b9e882cd4a124