Revert "Create_cow: Hash of target blocks should not be stored."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/8bfea28f-e9cc-4fc5-a223-2e5d41123b35). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id842c7de9d6e10b352856331d00ac61732e0608b