Revert "Create_cow: Hash of target blocks should not be stored."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b8a4c816-a5fb-4d10-9e2d-dff974b16139). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I2e72205793138a53d3dd15e9d701aae435ac0e04