Battery with question mark shows up after attach charger
When battery is not completely drained(remain battery level
is bigger than 0 before it shutdown, if attach charger bring
device to off-mode charge, the picture of battery with
question mark could show up before it display actually
battery level. This is due to the checking for time expire
is based on both battery level is 0 and battery status is
unknown, but the display of the picture just check the
battery status.

Bug: 300166399
Test: verify the battery with question no longer shows up
Change-Id: I455ffa88524de6730697a74946427dd2a26cd9f0