Revert "Only skip _16k/_64k kernel modules dirs on 4K builds"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2f485d63-7d42-45d8-a07a-030807a136ae). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I1434db8d96ad84242016d4ba3a3f618e7324d7dd