Revert "Add an include"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/01a7b760-d011-4143-b795-eaa62ce9b914). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I3f6d4eef28f78a03ff5312851a6085c28466f76b