Full uname data in sysprop
System properties are the simplest way of collecting system information.
Uname is a syscall that has no access restriction, and returns important
information about the current kernel.
Extend the ro system properties with all the output of uname to be later
used for device analysis, for example by a test report sent for device
certification.
This data will be later used to match kernel information with a build
fingerprint.

This is the output sampled from a recent Cuttlefish:

$ adb shell getprop | grep "ro\.kernel\.version"
[ro.kernel.version]: [6.1]
[ro.kernel.version_machine]: [x86_64]
[ro.kernel.version_nodename]: [(none)]
[ro.kernel.version_release]: [6.1.25-android14-11-g34fde9ec08a3-ab10675345]
[ro.kernel.version_sysname]: [Linux]
[ro.kernel.version_version]: [#1 SMP PREEMPT Thu Aug 17 19:23:24 UTC 2023]

Test: m && adb shell getprop | grep "ro\.kernel\.version"
Bug: 307675049
Change-Id: I15dc948cf2328a08276965a8b2d04e041565a7a5
Signed-off-by: Alessio Balsini <balsini@google.com>