libsnapshot: Implement CowWriterV3:EmitBlocks
Add in replace ops to WriterV3 without compression support (will be
added in later). CowReader also has to be changed since we no longer
have a CowFooter (our checks need to be updated correspondingly).
InitPost has to be implemented in this CL to make sure we write to the
correct position.

Test: cow_api_test
Change-Id: Ie979c72f842edd04337d900fd43dac8031207517