Distinguish read/write accesses in debuggerd reports.
Adds a feature that determines whether a given SIGSEGV (or other bad
memory access) was a read or write on aarch64, arm32, x86_64, x86, and
riscv.

The output happens next to the fault address, and looks like:

  signal 11 (SIGSEGV), code 2 (SEGV_ACCERR), fault addr
0x000071c21c95f000, read memory access

This is output in logcat, in a new field in the signal info in the
pbtombstone, and in the text tombstone as well.

Test: atest -a debuggerd_test
Test: Also manually verified using sanitizer-status on aarch64 and
x86_64 cuttlefish. I have not verified riscv, but I believe x86 and
arm32 should have been verified by the '-a' flag in atest.

Bug: N/A
Change-Id: Id0a15d858b1af5afa96b0d2b217997fc9ffec931