Revert "Reapply "Migrate the blkio controller to the v2 cgroup h..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f9017c15-7ae3-4ebc-9256-509223e82c6c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ifd234d380c7e6a43feded959264ac59f446b40a0