init: Fix a race condition between the stop and start actions
Bug 151795571 was fixed by replacing the 'restart' actions for the audio
HAL services into separate 'stop' and 'start' actions. That change
introduced a race condition. The 'start' action only starts an action if
the SVC_RUNNING flag is not set. The 'stop' action does not clear that
flag - that flag is only cleared after the init process has received the
SIGCHLD signal for the service process leader. Fix this race condition
by waiting until SIGCHLD has been received if a service is stopped.

Fixes: 308687042
Change-Id: I314b08dc6d3f1ceb9e713c1e860e56c65b3b6d2d
Signed-off-by: Bart Van Assche <bvanassche@google.com>