Revert "libsnapshot: Add Parser v3 + stub"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/18fa6d80-c538-4409-9037-10da8bd9a087). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I5bc134a4b5f358222ef2668426011b4a07c7b935