Revert "libsnapshot: Add Parser v3 + stub"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/143e42f2-0f4a-4f78-8d0a-05795e0efb8a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I8c16c802c00b9c990ec1783f494594a88a979b97