Revert "libsnapshot: Add Parser base class"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/143e42f2-0f4a-4f78-8d0a-05795e0efb8a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia27917ca67743bd40bad3509e8d8ca17c14dfe86