Revert "libsnapshot: Add Parser base class"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/687afc40-68ac-4bbe-b6eb-c5e3062d925f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Icd626f6bb5e0f0a6ca876542217a08d4847c1264