Revert "libsnapshot: Add Parser v3 + stub"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/687afc40-68ac-4bbe-b6eb-c5e3062d925f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I29071c7d866e289dd2d21617f8b57952026e9952