Revert "libsnapshot: Add Parser base class"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/0d0dc3ae-0cf1-4934-a88b-369da056e09b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I8a457a1c93ef1ea43d143d3a62d4601c3ba155d8