Revert "libsnapshot: Add Parser v3 + stub"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/0d0dc3ae-0cf1-4934-a88b-369da056e09b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I8280efdd34f50ff9a002b85fc92b51007a040675