Revert "libsnapshot: Add Parser base class"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f06f4922-5355-40a1-9e46-b65470812e7f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I8dec7d6f4a8cac136f0fac5e93482ffcd8e33780