Revert "libsnapshot: Add Parser v3 + stub"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f06f4922-5355-40a1-9e46-b65470812e7f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ib4fdee4b569dca2541c9d46431d8995fedb4a918