Revert "libsnapshot: Add Parser base class"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/8b392ca4-cbc4-4c55-9bdc-2514b40ffeb2). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I05e0cfec735a756c6d588cffe9e643331d50af1d