Revert "libsnapshot: Add Parser v3 + stub"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/8b392ca4-cbc4-4c55-9bdc-2514b40ffeb2). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia7cd7f49e6f684b826bb3615b7b600512449d26a