Revert "libsnapshot: Add Parser base class"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f8b6c980-6140-4770-9e74-2ca5db99e22d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ibe82037e147932c63be1d10889a0edb39a86ac9b