Revert "libsnapshot: Add Parser base class"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e4eda700-ff3f-4c49-be03-c915b45a433f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id7622b16c7f4e48dbef54e86801f3195f05b8788