Revert "libsnapshot: Add Parser v3 + stub"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e4eda700-ff3f-4c49-be03-c915b45a433f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia844177475ceda45e21d3fa335249ae61735be60