Revert "Reapply "Migrate the blkio controller to the v2 cgroup h..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/dfb05ed7-057a-4ecf-ba3a-de7b4edccee8). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I2bcd8e9a5c0d60c1db4fa57fe738455361dc86ea