Revert "Reapply "Migrate the blkio controller to the v2 cgroup h..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/52b11c59-ca70-46b8-a9a5-80838db3a516). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I29ab474b010a687514b64bf938eb9cb9ae219b60