Revert "Revert "Reapply "Migrate the blkio controller to the v2 ..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/76c4f492-2f4b-40de-801c-c3c282086c00). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I6b59054d3d58ab7fe22d12e2a616383248ecb495