Revert "libprocessgroup: Internal linkage for removeAllProcessGr..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/83de80dc-d23f-437e-932c-045ecb088fc4). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I8185b95e2a0792b0c8f40ec3b740efe71b086c9b