libprocessgroup: Kill processes even if the 'cgroup.procs' file is not found
If the wrong uid argument is passed when killing processes, no processes
are killed. Change this behavior into killing the process group
associated with the initial process ID. This is sufficient to make the
AudioCoreConfigTest pass with the blkio controller activated in the v2
cgroup hierarchy. Without this change that test fails systematically.

Bug: 309674654
Change-Id: Ida613cf53ec5aaff441e81c13ec27421dbb21133
Signed-off-by: Bart Van Assche <bvanassche@google.com>