add callstack check
when debuggerd recieved a signal, compare the callstack with historical
callstacks, if matches one, not crash. A property is added to control this feature.

when testing an APP in CI/CD, especially with MTE-enabled, APP crashes
at some callstacks which are known to developers and can be skipped.
Developers can enable this feature to find more bugs.

Change-Id: Ie896f3e23fd1ea74e762485472b24609f56bea4e
Signed-off-by: yaoping <yaoping@vivo.com>