libsnapshot: resume_point_count
We want to add a resume_point_count in the header to represent how many
resume points we've written. In the case that we've written less than
resume_buffer_size, we only want to read the valid resume points.

Also updating findopindex to use uint64_t to match the expected type of
label

Test: cow_api_test
Change-Id: I0a8971955439639f2d0f39d9d518c1145ae15c3d