Revert "Calculate ro.vendor.api_level with the new vendor API format"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/8b1d82cd-3488-465a-bdeb-7261347f682f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie92800f5eac4d5b24d40847b5297f29939170d14