Revert "Calculate ro.vendor.api_level with the new vendor API format"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/53bbcf06-5650-48f1-9a4c-2a9d2942528a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I8b0b88cffe7bcae636babbc751a22d11e3ee272e