Revert "libprocessgroup: Use cgroup.kill"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/91da3c52-9b76-498b-bdbd-a9de7d7ff53b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Icb8a59cc2041e6306fdf0976f7f8aabf3fdb6839