Revert "libprocessgroup: Poll on cgroup.events"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/91da3c52-9b76-498b-bdbd-a9de7d7ff53b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I0c2c6ab683088e5fe937cbfca18d62872e643f90