Revert "healthd: Add handling for "Calibration required" battery..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/9c4f8798-c2c1-47d6-9dae-e10308db9956). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I2d597f11eb228deb6506deb3e7e1b2051379052a