Revert "healthd: Add handling for "Calibration required" battery..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/9246d30a-4188-43d1-a7eb-67be3521c5de). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie25bacc9a76dd7a1c3003fead892500dde5efa34