Revert "healthd: Add handling for "Calibration required" battery..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/0453e5eb-1751-45aa-a8d5-3ccf70d5517d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If37bed295e645565aa7bc29f184ccda401f22f0a