Revert "healthd: Add handling for "Calibration required" battery..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/dceba43f-03ed-4802-8579-9aa10f91f4f1). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I6844528f65c1307643a70c216a81630f24498ccf