Revert "healthd: Add handling for "Calibration required" battery..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/7a6788d6-c0b1-4e6b-8d93-cc72a10bc450). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Idb0aa38e63c67661c6cc4ae60e63e3ef6107a858