Revert "healthd: Add handling for "Calibration required" battery..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/959ce9e3-33ba-4ebd-8c38-ef8ddb9c2b4d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I63319d427864bc90620b2d4737447dd2c005493a