Revert "healthd: Add handling for "Calibration required" battery..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/641664b4-2ec4-4926-8720-42105b8545ee). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I55a45adb92da0e600ac943fed606fa46a487d7fa