Revert "healthd: Add handling for "Calibration required" battery..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/0727d5bf-5b43-4fd4-ac93-0b1e9d99f940). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I8bdfd744bf2a4951a3e419831b13a44dcb69441f