Revert "healthd: Add handling for "Calibration required" battery..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/361429ea-0e45-4d7c-bf43-ce34ed6dd61f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie4e6cbb36752dfe275b39a8e86101a016c52d793