Revert "healthd: Add handling for "Calibration required" battery..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/807a5e72-20c7-4ce9-b774-643a2c2e2f38). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie6b78ca940a403882ef18caa8377d26ecd8ce7fe