Revert "healthd: Add handling for "Calibration required" battery..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/fb4a78e9-6485-4730-8b55-3b7bce6654f5). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I54014b977ee22f0303236ae6fd54bdb3c343fc59