Revert "remount: Detect when flashall has happened in the bootlo..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/8a605e2d-9765-4c52-b8a7-5a20594b972d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If83a8b0a2b5670d33a6ae42941b1ea3fae9f98ae