Revert "healthd: Add handling for "Calibration required" battery..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b531b5fe-3bd0-42af-81f6-5ad4eb3261a9). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic6c9bb79c17b67c3bcb57642350d7fa6224d38f6