adb-remount-test: Only test mounts that are remounted
Right now we assume all RW mounts (minus /data & special FS) to be
remounted by adb-remount and we apply the remount/overlayfs property
checks on them.
This would generate false positives when there is a partition that was
RW but not remounted by adb-remount. The test should only check mounts
that are remounted by adb-remount (transitioned from RO to RW after
adb-remount), and ignore those partitions that are already RW before
running adb-remount.

Bug: 313609600
Test: adb-remount-test
Change-Id: I94e8a35775271f557790a458781657eb3b24a6f5