libsnapshot: remove op count check
With variable block size compression being added, the number of ops
written cannot be calculated directly as easily since one op can cover
the data for multiple ops previously. We can get rid of this check as
within WriteOperation() we already make a check to see if we are
exceeding op_count_max limit.

Test: th
Change-Id: Ia55f152f5deb67a9022d0feff112345e72741dd3