Revert "TEST_MAPPING: don't run vts_libsnapshot_test in kernel-p..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/4d1a3e84-fd2b-4870-a8ee-67b9ed4f0774). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I1fba8c356ac49d74da81f7461ffce067446a93f8