Revert "libprocessgroup: Use pid_t consistently for TIDs"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/45a99abf-df3f-4469-b767-9e40e90ec41f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia165095ea034b095762511cc96186e6b47a04915