Revert "libprocessgroup: Use uid_t consistently for UIDs"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/45a99abf-df3f-4469-b767-9e40e90ec41f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id712eb6d938c9377c3757592540ffe5d0a577d68