Revert "libprocessgroup: Use pid_t consistently for PIDs"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/45a99abf-df3f-4469-b767-9e40e90ec41f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Icdc95b76dec754bb820b8ea9474f23ac21b0cea6