first_stage_console: Refactor RunScript()
Introduce ExecShell to DRY the 2 manual calls to execv, which supports a
variable number of arguments (by taking std:vector) and prints out a
more robust error message by generating it from its inputs. It uses PLOG
instead of logging errno manually (note that the return value of execv()
is guaranteed to always be -1) and FATAL to ensure that the process dies
(instead of manually calling _exit(2)) turning the following errors

[   10.779836][   T52] init: unable to execv /first_stage.sh, returned -1 errno 2
[   10.796387][   T51] init: unable to execv, returned -1 errno 2

into

[   10.789421][   T52] init: execv("/system/bin/sh", "/first_stage.sh") failed: No such file or directory
[   10.808625][   T51] init: execv("/system/bin/sh") failed: No such file or directory

Inline the call to RunScript() to hopefully make it simpler for readers
to track the lifetime of the various processes on different code paths.

Test: run first_stage_init
Change-Id: Ifaab2be032b2080a039209295d0b5a3759764ea7