init: snapuserd: Fix ranges for mlock()
It cannot be assumes that file  mappings in /proc/<pid>/maps will be
compeleted backed by the underlying file. [1]

Use MappedFileSize() to deduce the correct ranges for the mlock
calls when locking system pages in the OTA path.

While at it also clean up the some unreachable code (mlockall),
and improve error logging.

[1] SIGBUS at https://man7.org/linux/man-pages/man2/mmap.2.html#RETURN_VALUE

Test: Incremental OTA
Bug: 324952273
Change-Id: Ia2ab150e1b8de8c638f5b1acc1de83deb7ac5cff
Signed-off-by: Kalesh Singh <kaleshsingh@google.com>