Revert "snapshotctl: Add apply-update option"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/3346edbe-e590-4642-af59-1d35ff4acbe7). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic3321623a89052765b15b6f8e1d1b0baa6d38fe1