Revert "Define __INTRODUCED_IN_LLNDK"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2c3ff75d-5adb-421b-9b25-6a6c60c29e58). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I2c9c8de597fb820ebdb8c546e969799278999c97